# api/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Projeto EcosRev 🍀",
    "description": "➡️Documentação gerada automaticamente pelo módulo <a href='https://github.com/davibaltar/swagger-autogen' target='_blank'>swagger-autogen</a>."
  },
  "host": "localhost:3000",
  "basePath": "",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "access-token",
      "description": "Token de Acesso gerado após o login"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/hist/pontos": {
      "post": {
        "tags": [
          "Histórico"
        ],
        "summary": "POST para registrar pontos após leitura do QR code",
        "description": "Registra pontos no histórico do usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "idUser": {
                  "example": "any"
                },
                "points": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/hist/transacoes": {
      "post": {
        "tags": [
          "Histórico"
        ],
        "summary": "POST para registrar uma transação de benefício",
        "description": "Registra uma transação no histórico do usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idUser": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "points": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/hist/{idUser}": {
      "get": {
        "tags": [
          "Histórico"
        ],
        "summary": "GET histórico de pontos e transações de um usuário",
        "description": "Retorna histórico de pontos e transações de um usuário, filtrado por data",
        "parameters": [
          {
            "name": "idUser",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "start",
            "in": "query",
            "type": "string"
          },
          {
            "name": "end",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/hist/": {
      "get": {
        "tags": [
          "Histórico"
        ],
        "summary": "GET histórico de todos os usuários",
        "description": "Retorna o histórico de pontos e transações de todos os usuários, agrupado por idUser, filtrado por data",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "type": "string"
          },
          {
            "name": "end",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    }
  },
...
}